refactor folder structure add static folder add template folder add img folder add .gitignore create seperate css and js files for better maintainability rename main.py -> app.py move readme.jpg/logo.png to img folder remove unused imports remove logo route added document.title translation
