+++ static/styles.css
@@ -1,0 +1,36 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap");
+body {
+    font-family: "Poppins", sans-serif;
+    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+}
+.glass-morphism {
+    background: rgba(255, 255, 255, 0.25);
+    box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+    backdrop-filter: blur(4px);
+    -webkit-backdrop-filter: blur(4px);
+    border-radius: 10px;
+    border: 1px solid rgba(255, 255, 255, 0.18);
+}
+.fancy-table {
+    border-collapse: separate;
+    border-spacing: 0;
+    width: 100%;
+    border-radius: 10px;
+    overflow: hidden;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+.fancy-table th,
+.fancy-table td {
+    padding: 12px 15px;
+    text-align: left;
+}
+.fancy-table thead {
+    background-color: #8b5cf6;
+    color: white;
+}
+.fancy-table tbody tr:nth-child(even) {
+    background-color: rgba(255, 255, 255, 0.8);
+}
+.fancy-table tbody tr:nth-child(odd) {
+    background-color: rgba(255, 255, 255, 0.6);
+}
