+++ templates/index.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Çok Dilli Uçuş Takip Uygulaması</title>
+        <script src="https://cdn.tailwindcss.com"></script>
+        <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+        <script src="https://html2canvas.hertzen.com/dist/html2canvas.min.js"></script>
+        <link rel="stylesheet" href="./static/styles.css" />
+    </head>
+    <body class="min-h-screen py-6 flex flex-col justify-center sm:py-12">
+        <div class="relative py-3 sm:max-w-5xl sm:mx-auto">
+            <div class="glass-morphism relative px-4 py-10 shadow-lg sm:rounded-3xl sm:p-20">
+                <div class="max-w-5xl mx-auto">
+                    <div id="logo" class="w-32 h-32 mx-auto mb-4 bg-contain bg-no-repeat bg-center" style="background-image: url('./static/img/logo.png')"></div>
+                    <h1 id="mainTitle" class="text-3xl font-semibold mb-6 text-center text-white">Çok Dilli Uçuş Takip</h1>
+                    <div class="mb-4 text-center">
+                        <select id="languageSelect" class="px-4 py-2 rounded-md">
+                            <option value="en">🇬🇧 English</option>
+                            <option value="ru">🇷🇺 Русский</option>
+                            <option value="de">🇩🇪 Deutsch</option>
+                            <option value="tr">🇹🇷 Türkçe</option>
+                            <option value="es">🇪🇸 Español</option>
+                            <option value="fr">🇫🇷 Français</option>
+                        </select>
+                    </div>
+                    <p id="locationLabel" class="mb-4 text-center text-white"></p>
+                    <div class="mb-6">
+                        <input type="range" id="radiusSlider" min="10" max="250" value="100" class="w-full" />
+                        <p class="text-white text-center"><span id="radiusLabel"></span>: <span id="radiusValue">100</span> km</p>
+                    </div>
+                    <button id="refreshBtn" class="mb-4 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
+                        <span id="refreshLabel"></span>
+                    </button>
+                    <div class="overflow-x-auto mt-6">
+                        <table id="flightTable" class="fancy-table">
+                            <thead>
+                                <tr>
+                                    <th>ICAO24</th>
+                                    <th id="thCallsign"></th>
+                                    <th id="thCountry"></th>
+                                    <th id="thAltitude"></th>
+                                    <th id="thSpeed"></th>
+                                    <th id="thDirection"></th>
+                                    <th id="thLatitude"></th>
+                                    <th id="thLongitude"></th>
+                                </tr>
+                            </thead>
+                            <tbody></tbody>
+                        </table>
+                    </div>
+                    <p id="flightCount" class="mt-4 text-center text-white"></p>
+                    <button id="downloadBtn" class="mt-4 bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded">
+                        <span id="downloadLabel"></span>
+                    </button>
+                </div>
+            </div>
+        </div>
+        <script src="./static/scripts.js"></script>
+    </body>
+</html>
